+++ torchmetrics/audio/pit.py
@@ -18,7 +18,7 @@ from torch import Tensor, tensor
 
 from torchmetrics.functional.audio.pit import permutation_invariant_training
 from torchmetrics.metric import Metric
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning
 
 
 class PermutationInvariantTraining(Metric):
