+++ torchmetrics/classification/iou.py
@@ -17,7 +17,7 @@ import torch
 from deprecate import deprecated, void
 
 from torchmetrics.classification.jaccard import JaccardIndex
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning
 
 
 class IoU(JaccardIndex):
