+++ torchmetrics/classification/matthews_corrcoef.py
@@ -22,7 +22,7 @@ from torchmetrics.functional.classification.matthews_corrcoef import (
     _matthews_corrcoef_update,
 )
 from torchmetrics.metric import Metric
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning
 
 
 class MatthewsCorrCoef(Metric):
