+++ torchmetrics/functional/image/psnr.py
@@ -17,8 +17,7 @@ import torch
 from deprecate import deprecated, void
 from torch import Tensor, tensor
 
-from torchmetrics.utilities import rank_zero_warn, reduce
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning, rank_zero_warn, reduce
 
 
 def _psnr_compute(
