+++ torchmetrics/functional/image/ssim.py
@@ -19,9 +19,9 @@ from torch import Tensor
 from torch.nn import functional as F
 from typing_extensions import Literal
 
+from torchmetrics.utilities import _future_warning
 from torchmetrics.utilities.checks import _check_same_shape
 from torchmetrics.utilities.distributed import reduce
-from torchmetrics.utilities.warn import _future_warning
 
 
 def _gaussian(kernel_size: int, sigma: float, dtype: torch.dtype, device: torch.device) -> Tensor:
