+++ torchmetrics/functional/pairwise/manhatten.py
@@ -17,7 +17,7 @@ from deprecate import deprecated, void
 from torch import Tensor
 
 from torchmetrics.functional.pairwise.manhattan import pairwise_manhattan_distance
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning
 
 
 @deprecated(target=pairwise_manhattan_distance, deprecated_in="0.7", remove_in="0.8", stream=_future_warning)
