+++ torchmetrics/functional/text/cer.py
@@ -19,7 +19,7 @@ from deprecate import deprecated
 from torch import Tensor, tensor
 
 from torchmetrics.functional.text.helper import _edit_distance
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning
 
 
 def _cer_update(
