+++ torchmetrics/image/psnr.py
@@ -19,8 +19,7 @@ from torch import Tensor, tensor
 
 from torchmetrics.functional.image.psnr import _psnr_compute, _psnr_update
 from torchmetrics.metric import Metric
-from torchmetrics.utilities import rank_zero_warn
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning, rank_zero_warn
 
 
 class PeakSignalNoiseRatio(Metric):
