+++ torchmetrics/text/bert.py
@@ -20,8 +20,8 @@ from deprecate import deprecated
 from torchmetrics.functional import bert_score
 from torchmetrics.functional.text.bert import _preprocess_text
 from torchmetrics.metric import Metric
+from torchmetrics.utilities import _future_warning
 from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
-from torchmetrics.utilities.warn import _future_warning
 
 if _TRANSFORMERS_AVAILABLE:
     from transformers import AutoTokenizer
