+++ torchmetrics/text/cer.py
@@ -20,7 +20,7 @@ from torch import Tensor, tensor
 
 from torchmetrics.functional.text.cer import _cer_compute, _cer_update
 from torchmetrics.metric import Metric
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning
 
 
 class CharErrorRate(Metric):
