+++ torchmetrics/text/sacre_bleu.py
@@ -26,8 +26,8 @@ from typing_extensions import Literal
 from torchmetrics.functional.text.bleu import _bleu_score_update
 from torchmetrics.functional.text.sacre_bleu import _SacreBLEUTokenizer
 from torchmetrics.text.bleu import BLEUScore
+from torchmetrics.utilities import _future_warning
 from torchmetrics.utilities.imports import _REGEX_AVAILABLE
-from torchmetrics.utilities.warn import _future_warning
 
 AVAILABLE_TOKENIZERS = ("none", "13a", "zh", "intl", "char")
