+++ torchmetrics/text/wer.py
@@ -19,7 +19,7 @@ from torch import Tensor, tensor
 
 from torchmetrics.functional.text.wer import _wer_compute, _wer_update
 from torchmetrics.metric import Metric
-from torchmetrics.utilities.warn import _future_warning
+from torchmetrics.utilities import _future_warning
 
 
 class WordErrorRate(Metric):
