+++ torchmetrics/utilities/__init__.py
@@ -1,4 +1,3 @@
 from torchmetrics.utilities.data import apply_to_collection  # noqa: F401
 from torchmetrics.utilities.distributed import class_reduce, reduce  # noqa: F401
-from torchmetrics.utilities.prints import rank_zero_debug, rank_zero_info, rank_zero_warn  # noqa: F401
-from torchmetrics.utilities.warn import _future_warning  # noqa: F401
+from torchmetrics.utilities.prints import _future_warning, rank_zero_debug, rank_zero_info, rank_zero_warn  # noqa: F401
