+++ torchmetrics/utilities/prints.py
@@ -13,8 +13,10 @@
 # limitations under the License.
 import os
 import warnings
-from functools import wraps
+from functools import partial, wraps
 from typing import Any, Callable
+
+from _warnings import warn
 
 from torchmetrics import _logger as log
 
@@ -47,3 +49,4 @@ def _debug(*args: Any, **kwargs: Any) -> None:
 rank_zero_debug = rank_zero_only(_debug)
 rank_zero_info = rank_zero_only(_info)
 rank_zero_warn = rank_zero_only(_warn)
+_future_warning = partial(warn, category=FutureWarning)
